+++ stapi-pydantic/CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 ### Added
 
+- python `3.11` support ([#73](https://github.com/stapi-spec/pystapi/pull/73))
 - `stapi_type` and `stapi_version` ([#54](https://github.com/stapi-spec/pystapi/pull/54))
 
 ## [0.0.2] - 2025-04-02