+++ stapi-pydantic/CHANGELOG.md
@@ -6,6 +6,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 ## [Unreleased]
 
+### Added
+
+- python `3.11` support ([#73](https://github.com/stapi-spec/pystapi/pull/73))
+- `stapi_type` and `stapi_version` ([#54](https://github.com/stapi-spec/pystapi/pull/54))
+
+### Changed
+
+- `s/canceled/cancelled/` ([#75](https://github.com/stapi-spec/pystapi/pull/75))
+
+## [0.0.2] - 2025-04-02
+
 ### Changed
 
 - Added more top-level imports, removed conformance urls ([#51](https://github.com/stapi-spec/pystapi/pull/51))
@@ -14,4 +25,6 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 Initial release.
 
-[0.0.1]: https://github.com/stapi-spec/pystapi/releases/tag/v0.0.1
+[unreleased]: https://github.com/stapi-spec/pystapi/compare/stac-pydantic/stapi-pydantic%2Fv0.0.2...main
+[0.0.2]: https://github.com/stapi-spec/pystapi/compare/stac-pydantic/stapi-pydantic%2Fv0.0.1...stapi-pydantic%2Fv0.0.2
+[0.0.1]: https://github.com/stapi-spec/pystapi/releases/tag/stapi-pydantic%2Fv0.0.1